+++ libs/cimit-service/src/main/java/uk/gov/di/ipv/core/library/cimit/domain/CiMitVc.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.core.library.domain.cimitvc;
+package uk.gov.di.ipv.core.library.cimit.domain;
 
 import lombok.Getter;
 import uk.gov.di.ipv.core.library.annotations.ExcludeFromGeneratedCoverageReport;
@@ -7,7 +7,7 @@ import java.util.List;
 
 @ExcludeFromGeneratedCoverageReport
 @Getter
-public class EvidenceItem {
-    private String type;
-    private List<ContraIndicator> contraIndicator;
+public class CiMitVc {
+    private List<String> type;
+    private List<EvidenceItem> evidence;
 }
