+++ libs/cimit-service/src/main/java/uk/gov/di/ipv/core/library/service/CiMitService.java
@@ -23,12 +23,12 @@ import uk.gov.di.ipv.core.library.cimit.exception.CiPutException;
 import uk.gov.di.ipv.core.library.cimit.exception.CiRetrievalException;
 import uk.gov.di.ipv.core.library.config.ConfigurationVariable;
 import uk.gov.di.ipv.core.library.config.EnvironmentVariable;
-import uk.gov.di.ipv.core.library.domain.ContraIndicators;
+import uk.gov.di.ipv.core.library.cimit.domain.ContraIndicators;
 import uk.gov.di.ipv.core.library.domain.ErrorResponse;
-import uk.gov.di.ipv.core.library.domain.cimitvc.CiMitJwt;
-import uk.gov.di.ipv.core.library.domain.cimitvc.CiMitVc;
-import uk.gov.di.ipv.core.library.domain.cimitvc.ContraIndicator;
-import uk.gov.di.ipv.core.library.domain.cimitvc.EvidenceItem;
+import uk.gov.di.ipv.core.library.cimit.domain.CiMitJwt;
+import uk.gov.di.ipv.core.library.cimit.domain.CiMitVc;
+import uk.gov.di.ipv.core.library.cimit.domain.ContraIndicator;
+import uk.gov.di.ipv.core.library.cimit.domain.EvidenceItem;
 import uk.gov.di.ipv.core.library.exceptions.VerifiableCredentialException;
 import uk.gov.di.ipv.core.library.verifiablecredential.validator.VerifiableCredentialJwtValidator;
